% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IneQuant.R
\name{gini}
\alias{gini}
\title{Calculate gini coefficient}
\usage{
gini(x, weights = rep(1, length(x)))
}
\description{
This functions calculates gini coefficient, with or without weights
}
\arguments{
  \item{x}{input data}
  \item{weights}{corresponding weights to data, default weight is 1 for each data point}
}
\value{
Gini Coefficient
}
\details{
}
