% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IneQuant.R
\name{palma}
\alias{palma}
\title{Calculate palma ratio}
\usage{
palma(x, weights = rep(1, length(x)), bot = 40, top = 10)
}
\description{
This functions calculates palma ratio, with or without weights, with customizable ranges for top and bottom income ranges

}
\arguments{
  \item{x}{input data}
  \item{weights}{corresponding weights to data, default weight is 1 for each data point}
  \item{bot}{percentage of poorest population used, default 40}
  \item{top}{percentage of richest population used, default 10}
}
\value{Palma Ratio}
\details{
}
