% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IneQuant.R
\name{TheilsT}
\alias{TheilsT}
\title{Calculate Theil's T Coefficent}
\usage{
TheilsT(x, weights = rep(1, length(x)))
}
\description{
This functions calculates Theil's T coefficient, with or without weights
}
\arguments{
  \item{x}{input data}
  \item{weights}{corresponding weights to data, default weight is 1 for each data point}
}
\value{
Theil's T Coefficient
}
\details{
}
